add pre-shared key generation
This change does not add the pre-shared keys to the database but from
a security perspective this should not be done anyway.

If private keys exist the configuration is not overwritten so re-running
`genconfig` does not overwrite existing configuration.

TODO: add an option to generate new configuration (keys should be regularly
      rotated)

---

* pre-shared keys are recommended to help towards quantum resistance

  - https://www.wireguard.com/known-limitations/#post-quantum-secrecy

* this commit is based on:

  - https://github.com/k4yt3x/wg-meshconf/issues/10

  (with the missing json import added)